chore: 타입, 변수명 등 수정

+++ gagunokuga/gagunokuga_back/src/main/java/com/example/gagunokuga_back/furniture/service/FurnitureServiceImpl.java
@@ -42,7 +42,7 @@ public class FurnitureServiceImpl implements FurnitureService {
                             .build());
         }
         return FurnitureListResponse.builder()
-                .totalPages((long) furniturePage.getTotalPages())
+                .totalPages(furniturePage.getTotalPages())
                 .furnitures(furnitures)
                 .build();
     }
